requirements.txt
Flask==2.0.1
Werkzeug==2.0.3
flask-cors==3.0.10
google-cloud-storage==2.9.0
gunicorn==20.1.0
GDAL==3.4.1
numpy>=1.19.0 